+++ android/src/main/java/com/barkoder_flutter/Util.java
@@ -16,7 +16,7 @@ import java.io.ByteArrayOutputStream;
 class Util {
     private static final String TAG = Util.class.getSimpleName();
 
-    static String barkoderResultsToJsonString(Barkoder.Result[] results, Bitmap resultImage) {
+    static String barkoderResultsToJsonString(Barkoder.Result[] results, Bitmap[] thumbnails, Bitmap resultImage) {
         JSONObject resultJson = new JSONObject();
 
         try {
@@ -38,6 +38,33 @@ class Util {
 
             if (resultImage != null)
                 resultJson.put("resultImageAsBase64", bitmapImageToBase64(resultImage));
+
+            if (thumbnails[0] != null)
+                resultJson.put("resultThumbnailAsBase64", bitmapImageToBase64(thumbnails[0]));
+
+            if (results[0].images != null) {
+                for (Barkoder.BKImageDescriptor image : results[0].images) {
+                    if (image != null && image.image != null) {
+                        switch (image.name) {
+                            case "main":
+                                resultJson.put("mainImageAsBase64", bitmapImageToBase64(image.image));
+                                break;
+                            case "document":
+                                resultJson.put("documentImageAsBase64", bitmapImageToBase64(image.image));
+                                break;
+                            case "signature":
+                                resultJson.put("signatureImageAsBase64", bitmapImageToBase64(image.image));
+                                break;
+                            case "picture":
+                                resultJson.put("pictureImageAsBase64", bitmapImageToBase64(image.image));
+                                break;
+                            default:
+                                break;
+                        }
+                    }
+                }
+            }
+
         } catch (JSONException ex) {
             BarkoderLog.d(TAG, ex.getMessage());
         }
@@ -120,6 +147,8 @@ class Util {
                 return decoderConfig.Telepen;
             case Dotcode:
                 return decoderConfig.Dotcode;
+            case IDDocument:
+                return decoderConfig.IDDocument;
         }
         return null;
     }